backend-dev/requirements.txt
alembic==1.7.3
altgraph==0.17.3
anyio==3.6.2
APScheduler==3.10.1
asgiref==3.4.1
attrs==23.1.0
bson==0.5.10
CacheControl==0.13.1
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.0.12
click==8.0.1
cryptography==42.0.0
defusedxml==0.7.1
ecdsa==0.18.0
et-xmlfile==1.1.0
Faker==8.14.0
fastapi==0.95.1
firebase-admin==6.3.0
fonttools==4.39.4
fpdf2==2.7.4
gevent==23.9.1
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.12.0
httplib2==0.22.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
jsonify==0.5
Mako==1.1.5
MarkupSafe==2.0.1
msgpack==1.0.7
numpy==1.26.3
openpyxl==3.1.2
packaging==23.1
pandas==2.1.4
passlib==1.7.4
Pillow==9.5.0
pluggy==1.0.0
proto-plus==1.23.0
protobuf==4.25.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyJWT==2.8.0
PyMySQL==1.0.2
pyparsing==3.1.1
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.19.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
requests==2.26.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.25
starlette==0.26.1
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.5.0
tzdata==2023.4
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.15.0
uvloop==0.17.0
watchgod==0.7
websocket==0.2.1
websocket-client==1.7.0
websockets==10.0
XlsxWriter==3.1.0
zope.event==5.0
zope.interface==6.1
pypng==0.20220715.0
qrcode==7.4.2